cove: Move FIRFilter to libvectorizer
The FIRFilter class has no UI or QWidget depedencies.
