+++ 3rd-party/cove/libvectorizer/FIRFilter.h
